@@ -106,10 +119,10 @@
     "ts-loader": "^9.4.4",
     "typescript": "~5.5.3",
     "url": "^0.11.1",
-    "vue": "^2.7.15",
+    "vue": "^3.4.29",
     "vue-jest": "^3.0.7",
-    "vue-loader": "^15",
-    "vue-template-compiler": "^2.7.15",
+    "vue-loader": "^17.4.2",
+    "vue-style-loader": "~4.1.3",
     "webpack": "^5.75.0",
     "webpack-bundle-analyzer": "^4.7.0",
     "webpack-cli": "^5.0.0"
