@@ -56,11 +56,11 @@ export default {
       this.tooltipHook = this.$el.childNodes[0]
     }
 
-    const value = this.$slots.popover?[0] : this.text
+    const value = this.$slots.popover? this.$slots.popover()[0] : this.text
     initTooltip(this.tooltipHook, value, { place: this.placement, container: this.container })
   },
 
-  beforeDestroy () {
+  beforeUnmount () {
     destroyTooltip(this.tooltipHook)
   }
 }
