@@ -10,11 +10,11 @@
     @searched="emit($event, 'searched')"
     @selected="emit($event, 'selected')">
     <template
-      v-for="(_, scopedSlotName) in $scopedSlots"
-      v-slot:[scopedSlotName]="slotData">
+      v-for="(_, slotName) in $slots"
+      v-slot:[slotName]="slotData">
       <slot
-        :name="scopedSlotName"
-        v-bind="slotData" />
+        v-bind="slotData"
+        :name="slotName" />
     </template>
   </vue-omnibox>
 </template>
