@@ -6,7 +6,7 @@
     :class="{ 'u-pr-0' : isLast }"
     :data-col-field="headerField.field"
     :data-col-idx="idx">
-    <slot/>
+    <slot />
     <dp-resize-handle
       v-if="!isLast"
       :display-icon="isResizableColumn"
