@@ -27,27 +27,22 @@
         :checked="checked" />
     </th>
     <template v-for="(hf, idx) in headerFields">
-      <dp-resizable-column
-        v-if="isResizable"
-        :is-last="headerFields.length === idx"
+      <component
+        :is="isResizable ? DpResizableColumn : 'th'"
+        :is-last="headerFields.length === idx ? true : null"
         :header-field="hf"
         :next-header="headerFields[idx + 1]"
         :idx="idx">
         <slot
+          v-if="$slots[`header-${hf.field}`]()[0].children.length > 0"
           :name="`header-${hf.field}`"
           v-bind="hf">
           <span v-if="hf.label" v-text="hf.label" />
         </slot>
-      </dp-resizable-column>
-      <th
-        v-else
-        scope="col">
-        <slot
-          :name="`header-${hf.field}`"
-          v-bind="hf">
-          <span v-if="hf.label" v-text="hf.label" />
-        </slot>
-      </th>
+        <span
+          v-else-if="hf.label"
+          v-text="hf.label" />
+      </component>
     </template>
     <th
       v-if="hasFlyout"
