@@ -15,7 +15,7 @@
 </template>
 
 <script>
-import { nodeViewProps, NodeViewWrapper } from '@tiptap/vue-2'
+import { nodeViewProps, NodeViewWrapper } from '@tiptap/vue-3'
 import { de } from '../shared/translations'
 import { v4 as uuid } from 'uuid'
 
