@@ -4,62 +4,60 @@
     content-classes="w-fit"
     data-cy="editor:uploadModal"
     @modal:toggled="(isOpen) => { if (!isOpen) reset() }">
-    <template>
-      <h3
-        v-if="editAltTextOnly"
-        class="u-mb">
-        {{ translations.editImage }}
-      </h3>
-      <h3
-        v-else
-        class="u-mb">
-        {{ translations.insertImage }}
-      </h3>
-      <div v-if="!editAltTextOnly">
-        <dp-upload-files
-          allowed-file-types="img"
-          :basic-auth="basicAuth"
-          id="imageFile"
-          :get-file-by-hash="getFileByHash"
-          :max-file-size="20 * 1024 * 1024/* 20 MiB */"
-          :max-number-of-files="1"
-          ref="uploader"
-          :translations="{ dropHereOr: translations.uploadImage('20MB')}"
-          :tus-endpoint="tusEndpoint"
-          @upload-success="setFile" />
-      </div>
-      <div v-else>
-        <img
-          :alt="altText"
-          class="mb-4"
-          :src="this.imgSrc">
-      </div>
-      <dp-input
-        id="altText"
-        v-model="altText"
-        class="u-mb"
-        data-cy="uploadModal:altText"
-        :label="{
-          hint: translations.altTextHint,
-          text: translations.altText,
-        }" />
-      <div class="u-mt text-right w-full space-inline-s">
-        <button
-          class="btn btn--primary"
-          data-cy="uploadModal:save"
-          type="button"
-          @click="emitAndClose()"
-          v-text="editAltTextOnly ? Translator.trans('save') : Translator.trans('insert')">
-        </button>
-        <button
-          class="btn btn--secondary"
-          data-cy="uploadModal:abort"
-          type="button"
-          @click="closeAndReset()">
-          {{ Translator.trans('abort') }}
-        </button>
-      </div>
-    </template>
+    <h3
+      v-if="editAltTextOnly"
+      class="u-mb">
+      {{ translations.editImage }}
+    </h3>
+    <h3
+      v-else
+      class="u-mb">
+      {{ translations.insertImage }}
+    </h3>
+    <div v-if="!editAltTextOnly">
+      <dp-upload-files
+        allowed-file-types="img"
+        :basic-auth="basicAuth"
+        id="imageFile"
+        :get-file-by-hash="getFileByHash"
+        :max-file-size="20 * 1024 * 1024/* 20 MiB */"
+        :max-number-of-files="1"
+        ref="uploader"
+        :translations="{ dropHereOr: translations.uploadImage('20MB')}"
+        :tus-endpoint="tusEndpoint"
+        @upload-success="setFile" />
+    </div>
+    <div v-else>
+      <img
+        :alt="altText"
+        class="mb-4"
+        :src="this.imgSrc">
+    </div>
+    <dp-input
+      id="altText"
+      v-model="altText"
+      class="u-mb"
+      data-cy="uploadModal:altText"
+      :label="{
+        hint: translations.altTextHint,
+        text: translations.altText,
+      }" />
+    <div class="u-mt text-right w-full space-inline-s">
+      <button
+        class="btn btn--primary"
+        data-cy="uploadModal:save"
+        type="button"
+        @click="emitAndClose()"
+        v-text="editAltTextOnly ? Translator.trans('save') : Translator.trans('insert')">
+      </button>
+      <button
+        class="btn btn--secondary"
+        data-cy="uploadModal:abort"
+        type="button"
+        @click="closeAndReset()">
+        {{ Translator.trans('abort') }}
+      </button>
+    </div>
   </dp-modal>
 </template>
 
