@@ -1,5 +1,5 @@
 <template>
-  <div :class="containerWidth !== '' ? prefixClass(containerWidth) : false">
+  <div :class="containerClasses">
     <dp-label
       v-if="label.text !== ''"
       v-bind="{
@@ -7,8 +7,8 @@
         for: id,
         hint: labelHint,
         required: required
-      }" /><!--
- --><input
+      }" />
+    <input
       :id="id"
       :name="name !== '' ? name : null"
       :class="prefixClass(classes)"
@@ -17,7 +17,7 @@
       :data-dp-validate-error-fieldname="dataDpValidateErrorFieldname || label.text || null"
       :data-dp-validate-if="dataDpValidateIf !== '' ? dataDpValidateIf : null"
       :data-dp-validate-should-equal="dataDpValidateShouldEqual !== '' ? dataDpValidateShouldEqual : null"
-      :data-cy="dataCy !== '' ? dataCy : false"
+      :data-cy="dataCy !== '' ? dataCy : null"
       :aria-labelledby="ariaLabelledby"
       :maxlength="maxlength !== '' ? maxlength : null"
       :minlength="minlength !== '' ? minlength : null"
@@ -37,248 +37,254 @@
   </div>
 </template>
 
-<script>
-import { exactlengthHint, maxlengthHint, minlengthHint } from '~/utils'
+<script setup lang="ts">
+import { computed, ref, toRefs, watch } from 'vue'
+import { exactlengthHint, maxlengthHint, minlengthHint, prefixClass } from '~/utils'
 import DpLabel from '~/components/DpLabel'
 import { length } from '~/shared'
-import { prefixClassMixin } from '~/mixins'
 
-export default {
-  name: 'DpInput',
-
-  components: {
-    DpLabel
+const props = defineProps({
+  /**
+   * Reference another element on the page to define an accessible name if there is no label or
+   * you want to override the label.
+   */
+  ariaLabelledby: {
+    type: [String, null],
+    required: false,
+    default: null
   },
 
-  mixins: [prefixClassMixin],
+  /**
+   * Tell the browser if autocomplete is allowed or not. If enabled the browser is allowed
+   * to automatically complete the input. You can also provide a type of data which is expected.
+   */
+  autocomplete: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-  props: {
-    /**
-     * Reference another element on the page to define an accessible name if there is no label or
-     * you want to override the label.
-     */
-    ariaLabelledby: {
-      type: [String, Boolean],
-      required: false,
-      default: false
-    },
+  dataCounter: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-    /**
-     * Tell the browser if autocomplete is allowed or not. If enabled the browser is allowed
-     * to automatically complete the input. You can also provide a type of data which is expected.
-     */
-    autocomplete: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  dataCy: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-    dataCounter: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  dataDpValidateError: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-    dataCy: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  dataDpValidateErrorFieldname: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-    dataDpValidateError: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  dataDpValidateIf: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-    dataDpValidateErrorFieldname: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  dataDpValidateShouldEqual: {
+    type: String,
+    required: false,
+    default: ''
+  },
 
-    dataDpValidateIf: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  disabled: {
+    type: Boolean,
+    required: false,
+    default: null
+  },
 
-    dataDpValidateShouldEqual: {
-      type: String,
-      required: false,
-      default: ''
-    },
+  /**
+   * Use: when input field used with Icon || Button, then input field has padding right.
+   */
+  hasIcon: {
+    type: Boolean,
+    required: false,
+    default: false
+  },
 
-    disabled: {
-      type: Boolean,
-      required: false,
-      default: false
-    },
+  id: {
+    type: String,
+    required: true
+  },
 
-    /**
-     * Use: when input field used with Icon || Button, then input field has padding right.
-     */
-    hasIcon: {
-      type: Boolean,
-      required: false,
-      default: false
-    },
-
-    id: {
-      type: String,
-      required: true
-    },
-
-    label: {
-      type: Object,
-      default: () => ({
-        bold: true,
-        hide: false,
-        hint: '',
-        text: '',
-        tooltip: ''
-      }),
-      validator: (prop) => {
-        return Object.keys(prop).every(key => ['bold', 'hide', 'hint', 'text', 'tooltip'].includes(key))
-      }
-    },
-
-    /**
-     * Limit the maximum allowed number of characters to the given amount.
-     */
-    maxlength: length,
-
-    /**
-     * Define the minimum number of characters that need to be given.
-     */
-    minlength: length,
-
-    name: {
-      type: String,
-      required: false,
-      default: ''
-    },
-
-    pattern: {
-      type: String,
-      required: false,
-      default: ''
-    },
-
-    placeholder: {
-      type: String,
-      required: false,
-      default: ''
-    },
-
-    /**
-     * Set to false to prevent default behavior onEnter.
-     */
-    preventDefaultOnEnter: {
-      type: Boolean,
-      required: false,
-      default: true
-    },
-
-    readonly: {
-      type: Boolean,
-      required: false,
-      default: false
-    },
-
-    required: {
-      type: Boolean,
-      required: false,
-      default: false
-    },
-
-    /**
-     * When setting a number for the `size` prop, this is directly rendered
-     * as html attribute on the input element. Also, it is assumed that visual sizing
-     * based on that value shall be applied, that is why both container classes
-     * and element classes do not define any width styles when a size is set here.
-     */
-    size: {
-      type: [Number, null],
-      required: false,
-      default: null
-    },
-
-    type: {
-      type: String,
-      required: false,
-      default: 'text'
-    },
-
-    value: {
-      type: String,
-      required: false,
-      default: ''
-    },
-
-    /**
-     * Full width by default; set to 'auto' to have no width defined.
-     * @deprecated Apply width to the parent element of DpInput.
-     */
-    width: {
-      type: String,
-      default: 'u-1-of-1'
+  label: {
+    type: Object,
+    default: () => ({
+      bold: true,
+      hide: false,
+      hint: '',
+      text: '',
+      tooltip: ''
+    }),
+    validator: (prop) => {
+      return Object.keys(prop).every(key => ['bold', 'hide', 'hint', 'text', 'tooltip'].includes(key))
     }
   },
 
-  data () {
-    return {
-      currentValue: this.value
-    }
+  /**
+   * Limit the maximum allowed number of characters to the given amount.
+   */
+  maxlength: length,
+
+  /**
+   * Define the minimum number of characters that need to be given.
+   */
+  minlength: length,
+
+  name: {
+    type: String,
+    required: false,
+    default: ''
   },
 
-  computed: {
-    classes () {
-      let cssClasses = []
-      cssClasses = (this.size > 0) ? [] : [...cssClasses, 'u-1-of-1']
-      cssClasses = this.readonly || this.disabled ? [...cssClasses, 'bg-color--grey-light-2'] : cssClasses
-      cssClasses = this.type !== 'search' ? [...cssClasses, 'o-form__control-input'] : [...cssClasses, 'o-form__control-search']
-      cssClasses = this.hasIcon ? [...cssClasses, 'u-pr'] : cssClasses
-      return cssClasses.join(' ')
-    },
-
-    containerWidth () {
-      return (this.width === 'auto' || this.size > 0) ? '' : this.width
-    },
-
-    labelHint () {
-      const hint = typeof this.label.hint !== 'undefined' && this.label.hint !== '' ? [this.label.hint] : []
-
-      if (this.maxlength && !this.minlength) {
-        hint.push(maxlengthHint(this.currentValue.length, this.maxlength))
-      } else if (this.minlength && !this.maxlength) {
-        hint.push(minlengthHint(this.currentValue.length, this.minlength))
-      } else if (this.maxlength && this.minlength) {
-        if (this.maxlength === this.minlength) {
-          hint.push(exactlengthHint(this.currentValue.length, this.maxlength))
-        } else {
-          hint.push(maxlengthHint(this.currentValue.length, this.maxlength))
-          hint.push(minlengthHint(this.currentValue.length, this.minlength))
-        }
-      }
-
-      return hint
-    }
+  pattern: {
+    type: String,
+    required: false,
+    default: ''
   },
 
-  watch: {
-    value: function () {
-      this.currentValue = this.value
-    }
+  placeholder: {
+    type: String,
+    required: false,
+    default: ''
   },
 
-  methods: {
-    handleEnter (event) {
-      if (this.preventDefaultOnEnter === true) {
-        event.preventDefault()
-      }
-      this.$emit('enter')
+  /**
+   * Set to false to prevent default behavior onEnter.
+   */
+  preventDefaultOnEnter: {
+    type: Boolean,
+    required: false,
+    default: true
+  },
+
+  readonly: {
+    type: Boolean,
+    required: false,
+    default: null
+  },
+
+  required: {
+    type: Boolean,
+    required: false,
+    default: null
+  },
+
+  /**
+   * When setting a number for the `size` prop, this is directly rendered
+   * as html attribute on the input element. Also, it is assumed that visual sizing
+   * based on that value shall be applied, that is why both container classes
+   * and element classes do not define any width styles when a size is set here.
+   */
+  size: {
+    type: [Number, null],
+    required: false,
+    default: null
+  },
+
+  type: {
+    type: String,
+    required: false,
+    default: 'text'
+  },
+
+  value: {
+    type: String,
+    required: false,
+    default: ''
+  },
+
+  /**
+   * Full width by default; set to 'auto' to have no width defined.
+   * @deprecated Apply width to the parent element of DpInput.
+   */
+  width: {
+    type: String,
+    default: 'w-full'
+  }
+})
+
+const emit = defineEmits(['blur', 'enter', 'focus', 'input'])
+
+const { value, maxlength, minlength, width, size, label } = toRefs(props)
+const currentValue = ref(value.value)
+
+const classes = computed(() => {
+  let _classes: string[] = [
+   `px-1 py-0.5 max-w-full rounded-input
+    text-base leading-4 bg-surface
+    outline outline-1 outline-offset-0 outline-transparent
+    focus-visible:outline-interactive focus-visible:border-interactive focus-visible:z-above-zero
+    required:shadow-none`
+  ]
+
+  if (!(size.value && size.value > 0)) {
+    _classes.push('w-full')
+  }
+
+  if (props.readonly || props.disabled) {
+    _classes.push('bg-surface-light border-none cursor-default')
+  } else {
+    _classes.push('text-input bg-surface border border-input cursor-text')
+  }
+
+  if (props.hasIcon) {
+    _classes.push('pr-4')
+  }
+
+  return _classes.join(' ')
+})
+
+const containerClasses = computed(() => {
+  let _classes: string[] = [labelHint.value.length ? 'space-y-1' : 'space-y-0.5']
+
+  if (width.value !== 'auto' && size.value && size.value > 0) {
+    _classes.push(width.value)
+  }
+
+  return _classes
+})
+
+const labelHint = computed(() => {
+  const hint: string[] = typeof label.value.hint !== 'undefined' && label.value.hint !== '' ? [label.value.hint] : []
+
+  if (maxlength.value && !minlength.value) {
+    hint.push(maxlengthHint(currentValue.value.length, maxlength.value))
+  } else if (minlength.value && !maxlength.value) {
+    hint.push(minlengthHint(currentValue.value.length, minlength.value))
+  } else if (maxlength.value && minlength.value) {
+    if (maxlength.value === minlength.value) {
+      hint.push(exactlengthHint(currentValue.value.length, maxlength.value))
+    } else {
+      hint.push(maxlengthHint(currentValue.value.length, maxlength.value))
+      hint.push(minlengthHint(currentValue.value.length, minlength.value))
     }
   }
+
+  return hint
+})
+
+watch(value, (newVal) => {
+  currentValue.value = newVal
+})
+
+const handleEnter = (event: Event) => {
+  if (props.preventDefaultOnEnter === true) {
+    event.preventDefault()
+  }
+  emit('enter')
 }
 </script>
