@@ -1,14 +1,14 @@
 <template>
   <label
-    :class="prefixClass(['o-form__label flex', bold ? 'weight--bold' : 'weight--normal', hints.length > 0 ? 'has-hint' : '', hide ?? 'sr-only'])"
+    :class="classes"
     :for="labelFor">
     <span>
       <span v-cleanhtml="text" /><span v-if="required">*</span>
       <span
         v-if="hints.length > 0"
-        :class="prefixClass('block font-size-small weight--normal')">
+        :class="prefixClass('block text-sm font-normal')">
         <span
-          :class="prefixClass(['inline-block'])"
+          :class="prefixClass('inline-block')"
           :key="i"
           v-for="(h, i) in hints"
           v-cleanhtml="h" />
@@ -16,93 +16,83 @@
     </span>
     <dp-contextual-help
       v-if="tooltip !== ''"
-      :class="prefixClass('mt-px ml-0.5')"
+      :class="prefixClass('shrink-0 mt-px ml-0.5')"
       :text="tooltip" />
   </label>
 </template>
 
-<script>
-import { CleanHtml } from '~/directives'
-import { de } from '~/components/shared/translations'
-import { prefixClassMixin } from '~/mixins'
+<script setup lang="ts">
+import { computed } from 'vue'
+import { CleanHtml as vCleanhtml } from '~/directives' //NOSONAR see https://sonarsource.atlassian.net/browse/JS-97
+import { prefixClass } from '~/utils'
 import DpContextualHelp from '~/components/DpContextualHelp'
 
-export default {
-  name: 'DpLabel',
-  components: { DpContextualHelp },
-
-  directives: {
-    cleanhtml: CleanHtml
+const props = defineProps({
+  bold: {
+    type: Boolean,
+    required: false,
+    default: true
   },
 
-  mixins: [prefixClassMixin],
-
-  props: {
-    bold: {
-      type: Boolean,
-      required: false,
-      default: true
-    },
-
-    for: {
-      type: String,
-      required: true
-    },
-
-    // This is used to hide the label visually, but keep it accessible for screen readers
-    hide: {
-      type: Boolean,
-      required: false,
-      default: false
-    },
-
-    // Can be string or array (the second element being the "maxlength" hint).
-    hint: {
-      type: [String, Array],
-      required: false,
-      default: () => []
-    },
-
-    text: {
-      type: String,
-      required: true
-    },
-
-    tooltip: {
-      type: String,
-      required: false,
-      default: ''
-    },
-
-    required: {
-      type: Boolean,
-      required: false,
-      default: false
-    }
+  for: {
+    type: String,
+    required: true
   },
 
-  computed: {
-    /**
-     * List of Hints
-     *
-     * @return Array{String}
-     */
-    hints () {
-      if (this.hint) {
-        return this.wrapItemIntoArray(this.hint)
-      }
-      return []
-    },
-
-    labelFor () {
-      return this.for
-    }
+  // This is used to hide the label visually, but keep it accessible for screen readers
+  hide: {
+    type: Boolean,
+    required: false,
+    default: false
   },
 
-  methods: {
-    wrapItemIntoArray (item) {
-      return Array.isArray(item) ? item : [item]
-    }
+  // Can be string or array (the second element being the "maxlength" hint).
+  hint: {
+    type: [String, Array],
+    required: false,
+    default: () => []
+  },
+
+  text: {
+    type: String,
+    required: true
+  },
+
+  tooltip: {
+    type: String,
+    required: false,
+    default: ''
+  },
+
+  required: {
+    type: Boolean,
+    required: false,
+    default: false
   }
-}
+})
+
+const classes = computed(() => {
+  let cssClasses: string[] = ['flex']
+
+  if (props.hide) {
+    cssClasses.push('sr-only')
+  }
+
+  if (props.bold) {
+    cssClasses.push('font-semibold')
+  }
+
+  return cssClasses.map((selector) => prefixClass(selector))
+})
+
+const hints = computed(() => {
+  if (props.hint) {
+    return Array.isArray(props.hint) ? props.hint : [props.hint]
+  }
+
+  return []
+})
+
+const labelFor = computed(() => props.for)
+
 </script>
