@@ -12,7 +12,7 @@
       :data-cy="dataCy"
       :data-dp-validate-error-fieldname="dataDpValidateErrorFieldname || label.text || null"
       :required="required"
-      :name="name !== '' ? name : false"
+      :name="name !== '' ? name : null"
       class="o-form__control-select"
       :class="[disabled ? ' bg-color--grey-light-2' : '', classes]"
       :disabled="disabled"
