@@ -6,151 +6,135 @@
       <li
         v-for="(tab, idx) in tabs"
         :key="`tab:${idx}`"
-        :class="{ 'is-active': tab.isActive }"
-        style="margin-bottom: -1px;">
+        :class="{ 'is-active': activeTabId === tab.props.id }">
         <button
           role="tab"
           class="btn--blank o-link--default u-pb-0_5 border--bottom"
           :class="[
-            tab.isActive ? 'color-interactive border-interactive' : 'border--none color-text-muted',
+            activeTabId === tab.props.id ? 'color-interactive border-interactive' : 'border--none color-text-muted',
             tabSize === 'large' ? 'font-size-larger' : 'font-size-large'
           ]"
-          :aria-selected="tab.isActive"
-          :aria-controls="tab.id"
-          :data-cy="tab.id"
-          @click.prevent="handleTabClick(tab.id)">
-          {{ tab.label }}
+          :aria-selected="activeTabId === tab.props.id"
+          :aria-controls="tab.props.id"
+          :data-cy="tab.props.id"
+          @click.prevent="handleTabChange(tab.props.id)">
+          {{ tab.props.label }}
           <span
-            v-if="tab.suffix"
-            v-cleanhtml="tab.suffix" />
+            v-if="tab.props.suffix"
+            v-cleanhtml="tab.props.suffix" />
         </button>
       </li>
     </ul>
     <div>
-      <slot />
+      <slot :activeTabId="activeTabId" />
     </div>
   </div>
 </template>
 
-<script>
-import { CleanHtml } from '~/directives'
+<script setup>
+import { CleanHtml as vCleanhtml } from '~/directives'
+import { computed, onBeforeUnmount, onMounted, ref, useSlots } from 'vue'
 
-export default {
-  name: 'DpTabs',
-
-  directives: {
-    cleanhtml: CleanHtml
+const props = defineProps( {
+  activeId: {
+    type: String,
+    required: false,
+    default: null
   },
 
-  model: {
-    prop: 'activeId',
-    event: 'change'
+  tabSize: {
+    type: String,
+    required: false,
+    default: 'large',
+    validator: (prop) => ['medium', 'large'].includes(prop)
   },
 
-  props: {
-    activeId: {
-      type: String,
-      required: false,
-      default: null
-    },
+  /**
+   * Active tab state may be persisted via a url fragment. Also, all clicks are pushed
+   * to the browser history to enable state change when browser navigation is used.
+   */
+  useUrlFragment: {
+    type: Boolean,
+    required: false,
+    default: false
+  }
+})
 
-    tabSize: {
-      type: String,
-      required: false,
-      default: 'large',
-      validator: (prop) => ['medium', 'large'].includes(prop)
-    },
+const emit = defineEmits(['change'])
+const slots = useSlots()
 
-    /**
-     * Active tab state may be persisted via an Url fragment. Also, ab clicks are pushed
-     * to the browser history to enable state change when browser navigation is used.
-     */
-    useUrlFragment: {
-      type: Boolean,
-      required: false,
-      default: false
-    }
-  },
+const activeTabId = ref(null)
 
-  data: () => ({
-    tabs: []
-  }),
+const tabs = computed(() => {
+  return slots.default()[0].children.filter(child => child.type.name === 'DpTab')
+})
 
-  computed: {
-    tabIds () {
-      return this.tabs.map(tab => tab.id)
-    }
-  },
+/**
+ * When users click the back button, the tabs should behave accordingly.
+ */
+const handleHashChange = () => {
+  const hash = window.location.hash.slice(1)
 
-  methods: {
-    /**
-     * When users click the back button, the tabs should behave accordingly.
-     */
-    handleHashChange () {
-      const hash = window.location.hash.slice(1)
-
-      if (this.isTabId(hash)) {
-        this.setActiveTab(hash)
-      }
-    },
-
-    handleTabClick (id) {
-      this.$emit('change', id)
-      this.setActiveTab(id)
-
-      if (this.useUrlFragment) {
-        history.pushState(null, null, `#${id}`)
-      }
-    },
-
-    /**
-     * Check if a given string is a valid tab id.
-     * @param id
-     * @return {boolean}
-     */
-    isTabId (id) {
-      return this.tabs.map(tab => tab.id).includes(id)
-    },
-
-    setActiveTab (id) {
-      this.tabs.forEach(tab => {
-        tab.isActive = (tab.id === id)
-      })
-    },
-
-    /**
-     * If a hash specifies a tab, activate it. If a tab is activated
-     * via prop, activate that tab. Fallback to the first tab.
-     */
-    setInitialTab () {
-      const hash = window.location.hash.slice(1)
-
-      if (this.useUrlFragment && this.isTabId(hash)) {
-        this.setActiveTab(hash)
-      } else if (this.activeId) {
-        this.setActiveTab(this.activeId)
-      } else {
-        this.tabs[0].isActive = true
-      }
-    }
-  },
-
-  created () {
-    this.tabs = this.$children
-  },
-
-  mounted () {
-    this.setInitialTab()
-
-    if (this.useUrlFragment) {
-      window.addEventListener('hashchange', this.handleHashChange)
-    }
-  },
-
-  beforeDestroy () {
-    if (this.useUrlFragment) {
-      window.removeEventListener('hashchange', this.handleHashChange)
-    }
+  if (isTabId(hash)) {
+    setActiveTab(hash)
   }
 }
+
+const handleTabChange = (id) => {
+  emit('change', id)
+  setActiveTab(id)
+
+  if (props.useUrlFragment) {
+    history.pushState(null, null, `#${id}`)
+  }
+}
+
+/**
+ * Check if a given string is a valid tab id.
+ * @param id
+ * @return {boolean}
+ */
+const isTabId = (id) => {
+  return tabs.value.map(tab => tab.props.id).includes(id)
+}
+
+const setActiveTab = (id) => {
+  activeTabId.value = id
+}
+
+/**
+ * If a hash specifies a tab, activate it. If a tab is activated
+ * via prop, activate that tab. Fallback to the first tab.
+ */
+const setInitialTab = () => {
+  const hash = window.location.hash.slice(1)
+
+  if (props.useUrlFragment && isTabId(hash)) {
+    handleTabChange(hash)
+  } else if (props.activeId) {
+    handleTabChange(props.activeId)
+  } else {
+    handleTabChange(tabs.value[0].props.id)
+  }
+}
+
+onMounted (() => {
+  setInitialTab()
+
+  if (props.useUrlFragment) {
+    window.addEventListener('hashchange', handleHashChange)
+  }
+})
+
+onBeforeUnmount (() => {
+  if (props.useUrlFragment) {
+    window.removeEventListener('hashchange', handleHashChange)
+  }
+})
 </script>
+
+<style>
+li {
+  margin-bottom: -1px;
+}
+</style>
