@@ -13,6 +13,7 @@
       }" />
     <dp-upload
       :allowed-file-types="allowedFileTypes"
+      :allow-multiple-uploads="allowMultipleUploads"
       :basic-auth="basicAuth"
       :chunk-size="chunkSize"
       :class="[prefixClass('layout__item u-1-of-1-palm'), prefixClass(sideBySide ? 'u-1-of-2' : 'u-1-of-1')]"
@@ -39,7 +40,7 @@
       type="hidden"
       :name="name !== 'uploadedFiles' ? `uploadedFiles[${name}]` : 'uploadedFiles'"
       :required="required"
-      :data-dp-validate-if="dataDpValidateIf"
+      :data-dp-validate-if="dataDpValidateIf ? true : null"
       :value="fileHashes">
   </fieldset>
 </template>
@@ -76,6 +77,14 @@ export default {
       type: [Array, String],
       required: true,
       default: 'pdf'
+    },
+
+    /**
+     * Allow users to upload more files after uploading some
+     */
+    allowMultipleUploads: {
+      type: Boolean,
+      default: false
     },
 
     basicAuth: {
