@@ -1,13 +1,13 @@
 import validateMultiselect from './utils/validateMultiselect'
 
 const dpValidateMultiselectDirective = {
-  inserted (el, binding, vnode) {
+  mounted (el, binding) {
     if(!binding.value) {
       return
     }
 
     // Set initially correct validity value (no dropdown options => isValid)
-    const component = vnode.componentInstance
+    const component = binding.instance
     const hasOptions = component.groupValues ? component.options.some(group => group[component.groupValues].length > 0) : component.options.length > 0
     el.setAttribute('data-dp-validate-is-valid', !hasOptions)
     const validateMultiselectField = (e) => {
@@ -21,12 +21,12 @@ const dpValidateMultiselectDirective = {
     el.addEventListener('mouseup', validateMultiselectField)
   },
 
-  componentUpdated (el, binding, vnode) {
+  updated (el, binding) {
     if(!binding.value) {
       return
     }
 
-    const component = vnode.componentInstance
+    const component = binding.instance
     const hasOptions = component.groupValues ? component.options.some(group => group[component.groupValues].length > 0) : component.options.length > 0
     let isValid = checkValue(component.value)
     if (hasOptions === false) {
