@@ -4,10 +4,14 @@ const tailwindTheme = {
   borderRadius: require('./tokens/dist/tailwind/rounded'),
   boxShadow: require('./tokens/dist/tailwind/boxShadow'),
   fontSize: require('./tokens/dist/tailwind/fontSize'),
+  lineHeight: require('./tokens/dist/tailwind/space'),
   screens: require('./tokens/dist/tailwind/breakpoints'),
   spacing: require('./tokens/dist/tailwind/space'),
   zIndex: require('./tokens/dist/tailwind/zIndex'),
-  colors: require('./tokens/dist/tailwind/color')
+  colors: {
+    ...require('./tokens/dist/tailwind/color'),
+    'transparent': 'transparent'
+  }
 }
 
 const tailwindCorePluginsColor = ['backgroundColor', 'borderColor', 'textColor']
@@ -22,7 +26,16 @@ tailwindCorePluginsColor.forEach(corePlugin => {
 tailwindTheme.extend = {
   flexShrink: {
     2: '2'
-  }
+  },
+  animation: {
+    busy: 'busy 1.5s linear infinite',
+  },
+  keyframes: {
+    busy: {
+      'from': { 'background-position': '0 0' },
+      'to': { 'background-position': '30px 30px' },
+    },
+  },
 }
 
 module.exports = {
@@ -33,7 +46,13 @@ module.exports = {
     './.storybook/**/*.jsx'
   ],
   plugins: [
-    plugin(function({ addUtilities }) {
+    plugin(function({ addBase, addUtilities }) {
+      addBase({
+        html: {
+          'color': require('./tokens/dist/tailwind/textColor').default,
+        }
+      })
+
       addUtilities({
         /**
          * This overrides the Tailwind class of the same name with some
@@ -53,6 +72,19 @@ module.exports = {
           '&::-webkit-scrollbar': {
             'display': 'none'
           }
+        },
+        /**
+         * Visualize indeterminate system action. Use in conjunction with "animate-busy".
+         */
+        '.bg-busy': {
+          'background-image': 'linear-gradient(135deg, rgba(255, 255, 255, 0.3) 25%, transparent 25%, transparent 50%, rgba(255, 255, 255, 0.3) 50%, rgba(255, 255, 255, 0.3) 75%, transparent 75%, transparent)',
+          'background-size': '60px 60px'
+        },
+        /**
+         * Visualize invalid input elements. Applied by dpValidateMixin.
+         */
+        '.is-invalid': {
+          '@apply outline-interactive-warning border-interactive-warning': {},
         }
       })
     })
