@@ -5,7 +5,7 @@ import { shallowMount } from '@vue/test-utils'
 
 describe('should return result \'Auswahl aufheben\' when props are: checked = true, checkAll = true', () => {
   const wrapper = shallowMount(DpTreeListCheckbox, {
-    propsData: {
+    props: {
       checked: true,
       checkAll: true,
       name: 'someName'
@@ -19,7 +19,7 @@ describe('should return result \'Auswahl aufheben\' when props are: checked = tr
 
 describe('should return result aria.select.all when props are: checked = false, checkAll = true', () => {
   const wrapper = shallowMount(DpTreeListCheckbox, {
-    propsData: {
+    props: {
       checked: false,
       checkAll: true,
       name: 'someName'
@@ -33,7 +33,7 @@ describe('should return result aria.select.all when props are: checked = false, 
 
 describe('should return result \'Auswahl für Element aufheben\' when props are: checked = true, checkAll = false', () => {
   const wrapper = shallowMount(DpTreeListCheckbox, {
-    propsData: {
+    props: {
       checked: true,
       checkAll: false,
       name: 'someName'
@@ -47,7 +47,7 @@ describe('should return result \'Auswahl für Element aufheben\' when props are:
 
 describe('should return result \'Element auswählen\' when props are: checked = false, checkAll = false', () => {
   const wrapper = shallowMount(DpTreeListCheckbox, {
-    propsData: {
+    props: {
       checked: false,
       checkAll: false,
       name: 'someName'
