@@ -41,13 +41,17 @@ describe('DpTreeListNode', () => {
 
     wrapper = shallowMount(DpTreeListNode, {
       propsData,
-      mocks,
-      stubs
+      global: {
+        stubs,
+        mocks
+      }
     })
   })
 
   afterEach(() => {
-    wrapper.destroy()
+    if (wrapper) {
+      wrapper.unmount()
+    }
   })
 
   it('renders component correctly', () => {
@@ -93,7 +97,7 @@ describe('DpTreeListNode', () => {
 
   it('does not call the function setSelectionRecursively when the node has no children', () => {
     wrapper = shallowMount(DpTreeListNode, {
-      propsData: {
+      props: {
         children: [],
         draggable: true,
         handleChange: jest.fn(),
