@@ -5,7 +5,7 @@ import { shallowMount } from '@vue/test-utils'
 
 describe('DpCheckbox', () => {
   const wrapper = shallowMount(DpCheckbox, {
-    propsData: {
+    props: {
       id: 'checkboxId'
     }
   })
@@ -19,23 +19,27 @@ describe('DpCheckbox', () => {
 
   it('emits an event on change with the new `checked` state as argument', async () => {
     const componentWrapper = shallowMount(DpCheckbox, {
-      propsData: {
-        id: 'checkboxId'
+      props: {
+        id: 'checkboxId',
+        checked: true
       }
     })
 
-    const checkboxEl = componentWrapper.find('input[type="checkbox"]')
-    await checkboxEl.setChecked()
+    const checkboxEl = await componentWrapper.find('input[type="checkbox"]')
+    await componentWrapper.setProps({ checked: true })
+    await checkboxEl.trigger('change')
+    await componentWrapper.setProps({ checked: false })
+    await checkboxEl.trigger('change')
 
-    expect(componentWrapper.emitted().change[0][0]).toBeTruthy
-
-    await checkboxEl.setChecked(false)
-    expect(componentWrapper.emitted().change[1][0]).toEqual(false)
+    expect(componentWrapper.emitted()).toHaveProperty('change')
+    expect(componentWrapper.emitted()['change'][0][0]).toBe(true)
+    // One Event adds two element to the events array, therefore we have to check the 3rd value.
+    expect(componentWrapper.emitted()['change'][2][0]).toBe(false)
   })
 
   it('has the value of `valueToSend` as its value', async () => {
     const componentWrapper = shallowMount(DpCheckbox, {
-      propsData: {
+      props: {
         id: 'checkboxId'
       }
     })
